Stop inheriting from GraphQLDirective.
I had an epiphany last night that this class is not really a conceptual
subtype of `GraphQLDirective`, because a `GraphQLDirective` represents a
directive *declaration* in the schema, whereas `SchemaDirectiveVisitor`
defines behavior for any occurrences of the @directive elsewhere in the
schema. In particular, the `this.args` property of a `GraphQLDirective`
object represents the expected parameter types of the directive (and
default values, etc.), whereas `this.args` of a `SchemaDirectiveVisitor`
object represents the actual argument values passed to some specific
`@directive(...)` in the schema. Since there's no way to override the type
of the `args` property, and the subtyping relationship doesn't really make
sense anyway, I've removed `extends GraphQLDirective` from the class
declaration, in addition to renaming the class.